Flickr_ViewSets/Flickr/requirements.txt
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.5
djangorestframework==3.11.0
flickrapi==2.4.0
idna==2.9
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
oauthlib==3.1.0
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9